Move sponsors components
