+++ src/Features/Sponsors/Components/sponsorSubsection/index.js
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
-import { media } from '../../theme/media'
+import { media } from '../../../../theme/media'
 
 const BadgesContainer = styled.div`
   display: flex;
