+++ src/pages/sponsors.js
@@ -4,8 +4,8 @@ import styled from 'styled-components'
 import ReactMarkdown from 'react-markdown'
 import { Container } from '../components/grid/'
 import ImageBanner from '../components/imageBanner'
-import SponsorBadge from '../components/sponsorBadge'
-import SponsorsSubsection from '../components/sponsorSubsection'
+import SponsorBadge from '../Features/Sponsors/Components/sponsorBadge'
+import SponsorsSubsection from '../Features/Sponsors/Components/sponsorSubsection/index'
 import constants from '../constants'
 import theme from '../theme/theme'
 import { media } from '../theme/media'
